backend/youtube_lesson_planner/supabase_client.py
import os

from dotenv import load_dotenv
from langchain_community.vectorstores import SupabaseVectorStore
from langchain_openai import OpenAIEmbeddings
from supabase import create_client

load_dotenv()

supabase_url = os.environ.get("SUPABASE_URL")
supabase_key = os.environ.get("SUPABASE_KEY")
supabase = create_client(supabase_url, supabase_key)

embeddings = OpenAIEmbeddings(model="text-embedding-3-small", dimensions=1536)

vectorstore = SupabaseVectorStore(
    supabase,
    embedding=embeddings,
    table_name="transcripts",
    query_name="match_transcripts",
)
